examples/requirements.txt
# Core dependencies
memories-dev>=0.1.0
nltk>=3.8.1
numpy>=1.24.0
pandas>=2.0.0

# API integrations
requests>=2.31.0
aiohttp>=3.9.0

# Data processing
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0.1

# Optional - Specific to examples
earthengine-api>=0.1.37  # For water bodies monitoring
geopy>=2.4.0  # For location analysis
textblob>=0.17.1  # For sentiment analysis 